requirements.txt
# Core dependencies
torch>=2.0.0
torchaudio>=2.0.0
transformers>=4.30.0
openai-whisper>=20231117
faster-whisper>=0.9.0
nemo-toolkit[asr]>=1.20.0
nemo-toolkit[nlp]>=1.20.0

# Audio processing
librosa>=0.10.0
soundfile>=0.12.1
pydub>=0.25.1
demucs>=4.0.0

# Speaker diarization
pyannote.audio>=3.0.0
pyannote.core>=5.0.0

# Utilities
numpy>=1.24.0
scipy>=1.10.0
scikit-learn>=1.3.0
pandas>=2.0.0
tqdm>=4.65.0
click>=8.1.0
python-dotenv>=1.0.0

# API and web framework
fastapi>=0.100.0
uvicorn[standard]>=0.23.0
python-multipart>=0.0.6
aiofiles>=23.1.0

# Database and caching
redis>=4.6.0
sqlalchemy>=2.0.0
alembic>=1.11.0

# Monitoring and logging
prometheus-client>=0.17.0
structlog>=23.1.0

# Testing
pytest>=7.4.0
pytest-asyncio>=0.21.0
httpx>=0.24.0
